refactor: 대학 관련 파일 구조 개선
- `universities.py` -> `university.py`
- `db/models` 하위 파일명은 단수로 작성
- 건물 테이블 `university.py`에 통합
