+++ backend/db/models/university.py
@@ -1,8 +1,7 @@
-from typing import List
-from sqlalchemy import ForeignKey, Integer, String
-from sqlalchemy.orm import relationship, Mapped, mapped_column
+from sqlalchemy import Float, ForeignKey, Integer, String
+from sqlalchemy.orm import relationship, mapped_column
 from db.common import Base
-from buildings import BuildingModel
+
 
 class UniversityModel(Base):
     """단과대학 테이블"""
@@ -11,10 +10,7 @@ class UniversityModel(Base):
 
     id = mapped_column(Integer, primary_key=True, autoincrement=True)
     name = mapped_column(String, nullable=False)
-    departments: Mapped[List["DepartmentModel"]
-                        ] = relationship(back_populates="university")
-    buildings = relationship("BuildingModel", back_populates="university")
-    lectures = relationship("BuildingModel", back_populates="university")
+    departments = relationship("DepartmentModel", back_populates="university")
 
 
 class DepartmentModel(Base):
@@ -25,15 +21,15 @@ class DepartmentModel(Base):
     id = mapped_column(Integer, primary_key=True, autoincrement=True)
     university_id = mapped_column(ForeignKey("universities.id"))
     name = mapped_column(String, nullable=False)
-    university: Mapped["UniversityModel"] = relationship(
-        back_populates="departments"
-    )
+    university = relationship("UniversityModel", back_populates="departments")
 
     majors = relationship("MajorModel", back_populates="department")
     professors = relationship("ProfessorModel", back_populates="department")
     notices = relationship("NoticeModel", back_populates="department")
     buildings = relationship("BuildingsModel", back_populates="department")
-    lectures = relationship("LectureModel", back_populates="department")
+
+    subjects = relationship("SubjectModel", back_populates="department")
+    courses = relationship("CourseModel", back_populates="department")
 
 
 class MajorModel(Base):
@@ -48,3 +44,24 @@ class MajorModel(Base):
     department = relationship("DepartmentModel", back_populates="majors")
     professors = relationship("ProfessorModel", back_populates="major")
     buildings = relationship("BuildingsModel", back_populates="major")
+
+
+class BuildingModel(Base):
+    """대학 건물 테이블"""
+
+    __tablename__ = "buildings"
+
+    id = mapped_column(Integer, primary_key=True, autoincrement=True)
+    name = mapped_column(String, nullable=False, unique=True)
+    building_num = mapped_column(Integer, nullable=False, unique=True)
+
+    longitude = mapped_column(Float, nullable=False)
+    latitude = mapped_column(Float, nullable=False)
+
+    university_id = mapped_column(ForeignKey("universities.id"))
+    department_id = mapped_column(ForeignKey("departments.id"))
+
+    universities = relationship("UniversityModel", back_populates="buildings")
+    departments = relationship("DepartmentModel", back_populates="buildings")
+
+    timetables = relationship("CourseTimeTableModel", back_populates="building")
